% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplezoo-package.R
\docType{package}
\name{samplezoo-package}
\alias{samplezoo-package}
\title{samplezoo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
