% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplezoo-package.R
\docType{package}
\name{samplezoo-package}
\alias{samplezoo-package}
\title{samplezoo: Generate Samples with a Variety of Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simplifies the process of generating samples from various probability distributions, enabling users to create data frames quickly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nvietto.github.io/samplezoo}
  \item \url{https://github.com/nvietto/samplezoo}
  \item Report bugs at \url{https://github.com/nvietto/samplezoo/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
