% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplezoo-package.R
\docType{package}
\name{samplezoo-package}
\alias{samplezoo-package}
\title{samplezoo: Generate Samples with a Variety of Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nvietto.github.io/samplezoo/}
}

}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
