% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplezoo-package.R
\docType{package}
\name{samplezoo-package}
\alias{samplezoo-package}
\title{samplezoo: Generate Samples with a Variety of Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simplifies the process of generating samples from a variety of probability distributions, allowing users to quickly create data frames for demonstrations, troubleshooting, or teaching purposes. Data is available in multiple sizes—small, medium, and large. For more information, refer to the package documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nvietto/samplezoo}
  \item \url{https://nvietto.github.io/samplezoo/}
  \item Report bugs at \url{https://github.com/nvietto/samplezoo/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
