% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplezoo-package.R
\docType{package}
\name{samplezoo-package}
\alias{samplezoo-package}
\title{samplezoo: Generate Samples with a Variety of Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'samplezoo' package streamlines the process of generating samples from various probability distributions, enabling users to quickly create data frames for demonstrations, troubleshooting, or teaching. By prioritizing simplicity and efficiency,'samplezoo' reduces the need for repetitive code, making it particularly useful for beginners or anyone seeking to save time. The package implements standard methods for generating random samples from probability distributions commonly available in base R, with no specific external references. For more details, visit the package documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nvietto/samplezoo}
  \item \url{https://nvietto.github.io/samplezoo/}
  \item Report bugs at \url{https://github.com/nvietto/samplezoo/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
